renamed ENLSVG files to use snake_case
