--- src/extlibs/enlsvg/Pathfinding/Enlsvg.h
@@ -1,8 +1,0 @@
-#ifndef _ENLSVG_H_
-#define _ENLSVG_H_
-
-#include "EnlsvgAlgorithm.h"
-#include "Grid.h"
-#include "PathfindingDataTypes.h"
-
-#endif
