+++ src/extlibs/enlsvg/Pathfinding/enlsvg.h
@@ -1,0 +1,8 @@
+#ifndef _ENLSVG_H_
+#define _ENLSVG_H_
+
+#include "enlsvg_algorithm.h"
+#include "grid.h"
+#include "pathfinding_data_types.h"
+
+#endif
