+++ src/extlibs/enlsvg/Pathfinding/fraction.h
