+++ src/extlibs/enlsvg/Pathfinding/indirect_heap.h
