+++ src/extlibs/enlsvg/Pathfinding/pathfinding_data_types.h
