+++ src/extlibs/enlsvg/src/enlsvg_algorithm.cpp
@@ -1,11 +1,11 @@
-#include "extlibs/enlsvg/Pathfinding/EnlsvgAlgorithm.h"
+#include "extlibs/enlsvg/Pathfinding/enlsvg_algorithm.h"
 
 #include <algorithm>
 
-#include "extlibs/enlsvg/Pathfinding/EnlsvgGraph.h"
-#include "extlibs/enlsvg/Pathfinding/Grid.h"
-#include "extlibs/enlsvg/Pathfinding/IndirectHeap.h"
-#include "extlibs/enlsvg/Pathfinding/LineOfSightScanner.h"
+#include "extlibs/enlsvg/Pathfinding/enlsvg_graph.h"
+#include "extlibs/enlsvg/Pathfinding/grid.h"
+#include "extlibs/enlsvg/Pathfinding/indirect_heap.h"
+#include "extlibs/enlsvg/Pathfinding/line_of_sight_scanner.h"
 
 namespace Pathfinding
 {
