+++ src/extlibs/enlsvg/src/grid.cpp
@@ -1,4 +1,4 @@
-#include "extlibs/enlsvg/Pathfinding/Grid.h"
+#include "extlibs/enlsvg/Pathfinding/grid.h"
 
 #include <iostream>
