+++ src/software/ai/navigator/path_planner/enlsvg_path_planner.h
@@ -3,7 +3,7 @@
 #include <boost/functional/hash/hash.hpp>
 #include <queue>
 
-#include "extlibs/enlsvg/Pathfinding/Enlsvg.h"
+#include "extlibs/enlsvg/Pathfinding/enlsvg.h"
 #include "software/ai/navigator/obstacle/obstacle.hpp"
 #include "software/geom/linear_spline2d.h"
 #include "software/logger/logger.h"
